+++ tests/test_resources.py
@@ -361,7 +361,7 @@ def test_doc_from_responses():
             )
         ]
     )  # noqa: E501
-    x = _resources.doc_from_responses(responses, False)
+    x = _resources.doc_from_responses(responses, False, False)
     assert x == RESPONSE_DOC
 
 
@@ -446,7 +446,7 @@ def test_parse_params():
         "description": "nah!",
         "type": "integer",
     }
-    x, y = _resources.parse_params([param, param2], "summary!", "get")
+    x, y = _resources.parse_params([param, param2], "summary!", "get", "objects")
     expect_x = [
         {
             "in": "query",
@@ -466,6 +466,7 @@ def test_parse_params():
     ]
     expect_y = (
         "summary!\n\n"
+        "API URL: ``GET /objects``\n\n"
         "Parameters\n"
         "----------\n"
         "b : int\n"